Generic/Strongly typed List<T> conversion from arraylists in all XmlQuest/*
Another joucy addon is the RewardAction in "XmlQuestHolder", in this string an action could be done to the player when completing the quest with "AutoReward" to true, playing a song, damage, teleporting, changes. all with the same XmlSpawner syntax, considering the invoker/trigger mobile as the owner of the questbook.

Also a bugfix occurred during the screening of the code, questpoints where not correctly calculated for taking rewards and even if you remove the "CanSeeReward" from XmlQuestHolder, if the player had registered the macro or tried to sent a crafted packet with the correct button number, he could open the bag.
The bag under the same XmlQuestHolder now is more secure, it doesn't allow item stacking, thus preventing to the player to stack gold with the gold present, or put other items, that will anyway vanish when completed, in the same bag.